\name{SOP-package}
\alias{SOP-package}
\alias{SOP}
\docType{package}
\title{
\packageTitle{SOP}
}
\description{
\packageDescription{SOP}
}
\details{

\packageIndices{SOP}

This package incorporates the function \code{sop()} which enables users to estimate multidimensional generalised P-spline regression models with overlapping penalties. For a complete list of functions use \code{library(help = SOP)}.
}
\author{
\packageAuthor{SOP}

Maintainer: \packageMaintainer{SOP}
}
\references{
Eilers, P.H.C. and Marx, B.D. (1996). Flexible smoothing with B-splines and penalties. \emph{Statistical Science}, \bold{11 (2)}, 89--121.

Rodriguez-Alvarez, M.X., Lee, D. J., Kneib, T., Durban, M., and Eilers, P. (2015). Fast smoothing parameter separation in multidimensional generalized P-splines: the SAP algorithm. \emph{Statistics and Computing}, \bold{25 (5)}, 941--957.

Rodriguez-Alvarez, M.X., Durban, M., Lee, D. J. and Eilers, P. (2019). On the estimation of variance parameters in non-standard generalised linear mixed models: application to penalised smoothing. \emph{Statistics and Computing}, \bold{29 (3)}, 483--500.
}
\keyword{ package }
